Clean up logging around write_bash_code
This patch removes code that is not needed.

The parser_message is redundant now that the logging system writes the
error messages.

The error message for write_bash_code never worked: if writing fails,
write_bash_code() throws an exception which is not caught and the user
gets a stacktrace that is actually quite helpful.

Change-Id: I1e54d65bab21ee78928fd90b0246d33e66c560d3